Combine and deduplicate put and patch handlers
Combine the put and patch handlers into a single configure file to
allow for additional de-duplication of code.

Use the strategy pattern to handle the unique parts of put and patch.
